refactor: SocialLoginButton 컴포넌트 위치 이동 #11
